Tic Tak Toe evo neural net
I have successfully made an evolving AI that randomly changes values when wrong and creates a new default when it makes progress.

I need a way to save the AI after training it and making alternate AI paths. Also AI cheats but it still sucks so need to fix that. Finally something to be done about draws.